src/integrations/prefect-nomad/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "prefect-nomad"
description = "Prefect integrations with HashiCorp Nomad."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache License 2.0" }
keywords = ["prefect"]
authors = [{ name = "Prefect Technologies, Inc.", email = "help@prefect.io" }]
classifiers = [
  "Natural Language :: English",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries",
]
dependencies = ["prefect>=3.6.17", "python-nomad>=2.1.0"]
dynamic = ["version"]

[dependency-groups]
dev = [
  "aiohttp",
  "coverage",
  "interrogate",
  "mkdocs-gen-files",
  "mkdocs-material",
  "mkdocs",
  "mkdocstrings[python]",
  "mypy",
  "pillow",
  "pre-commit",
  "pytest-asyncio",
  "pytest >= 8.3",
  "pytest-asyncio>=0.25.3",
  "pytest-env",
  "pytest-xdist",
]

[project.urls]
Homepage = "https://github.com/PrefectHQ/prefect/tree/main/src/integrations/prefect-nomad"

[project.entry-points."prefect.collections"]
prefect_nomad = "prefect_nomad"

[tool.setuptools_scm]
version_file = "prefect_nomad/_version.py"
root = "../../.."
tag_regex = "^prefect-nomad-(?P<version>\\d+\\.\\d+\\.\\d+(?:[a-zA-Z0-9]+(?:\\.[a-zA-Z0-9]+)*)?)$"
fallback_version = "0.0.0"
git_describe_command = 'git describe --dirty --tags --long --match "prefect-nomad-*[0-9]*"'

[tool.interrogate]
ignore-init-module = true
ignore_init_method = true
exclude = ["prefect_nomad/_version.py", "tests"]
fail-under = 95
omit-covered-files = true

[tool.coverage.run]
omit = ["tests/*", "prefect_nomad/_version.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"
env = ["PREFECT_TEST_MODE=1"]

[tool.uv.sources]
prefect = { path = "../../../" }
